Move updates to own endpoint
