+++ ert_storage/endpoints/ensembles.py
@@ -35,41 +35,3 @@ def get_ensemble(*, db: Session = Depends(get_db), ensemble_id: int) -> js.Ensem
         children=[child.ensemble_result_id for child in ens.children],
         parent=ens.parent.ensemble_reference_id if ens.parent else None,
     )
-
-
-@router.post("/ensembles/{id}/updates", response_model=js.UpdateOut)
-def create_update(
-    *,
-    db: Session = Depends(get_db),
-    id: int,
-    update: js.UpdateIn,
-) -> js.UpdateOut:
-
-    update_obj = ds.Update(
-        algorithm=update.algorithm,
-        ensemble_reference_id=update.ensemble_reference_id,
-    )
-    db.add(update_obj)
-    transformations = (
-        update.observation_transformations if update.observation_transformations else []
-    )
-    observation_ids = [trans.observation_id for trans in transformations]
-    observations = (
-        db.query(ds.Observation).filter(ds.Observation.id.in_(observation_ids)).all()
-    )
-
-    observation_transformations = [
-        ds.ObservationTransformation(
-            active_list=observation_transformation.active,
-            scale_list=observation_transformation.scale,
-            observation=observation,
-            update=update_obj,
-        )
-        for observation_transformation, observation in zip(
-            transformations, observations
-        )
-    ]
-
-    db.add_all(observation_transformations)
-    db.commit()
-    return update_obj
